refactoring dao specifications
