+++ tracker-dao/src/test/java/org/example/tracker/dao/repository/EmployeeRepositoryTest.java
@@ -2,7 +2,7 @@ package org.example.tracker.dao.repository;
 
 import org.example.tracker.dao.entity.EmployeeEntity;
 import org.example.tracker.dao.repository.helper.EmployeeFilter;
-import org.example.tracker.dao.repository.specification.EmployeeSpecs;
+import org.example.tracker.dao.specification.EmployeeSpecs;
 import org.example.tracker.dto.employee.EmployeeFilterParam;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
